Refactor TransferData to remove unused fields in Contact table
