+++ core/src/main/java/google/registry/model/transfer/TransferData.java
@@ -22,11 +22,7 @@ import com.googlecode.objectify.annotation.IgnoreSave;
 import com.googlecode.objectify.condition.IfNull;
 import google.registry.model.Buildable;
 import google.registry.model.EppResource;
-import google.registry.model.billing.BillingEvent;
-import google.registry.model.domain.Period;
-import google.registry.model.domain.Period.Unit;
 import google.registry.model.eppcommon.Trid;
-import google.registry.model.poll.PollMessage;
 import google.registry.persistence.VKey;
 import google.registry.util.TypeUtils.TypeInstantiator;
 import java.util.Set;
@@ -37,7 +33,6 @@ import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.MappedSuperclass;
 import javax.persistence.Transient;
-import org.joda.time.DateTime;
 
 /**
  * Common transfer data for {@link EppResource} types. Only applies to domains and contacts; hosts
@@ -46,8 +41,6 @@ import org.joda.time.DateTime;
 @MappedSuperclass
 public abstract class TransferData<B extends TransferData.Builder<?, ?>> extends BaseTransferObject
     implements Buildable {
-
-  // public static final TransferData EMPTY = new TransferData();
 
   /** The transaction id of the most recent transfer request (or null if there never was one). */
   @Embedded
@@ -60,37 +53,6 @@ public abstract class TransferData<B extends TransferData.Builder<?, ?>> extends
         column = @Column(name = "transfer_client_txn_id"))
   })
   Trid transferRequestTrid;
-
-  /**
-   * The period to extend the registration upon completion of the transfer.
-   *
-   * <p>By default, domain transfers are for one year. This can be changed to zero by using the
-   * superuser EPP extension.
-   */
-  @Embedded
-  @AttributeOverrides({
-    @AttributeOverride(name = "unit", column = @Column(name = "transfer_renew_period_unit")),
-    @AttributeOverride(name = "value", column = @Column(name = "transfer_renew_period_value"))
-  })
-  Period transferPeriod = Period.create(1, Unit.YEARS);
-
-  /**
-   * The registration expiration time resulting from the approval - speculative or actual - of the
-   * most recent transfer request, applicable for domains only.
-   *
-   * <p>For pending transfers, this is the expiration time that will take effect under a projected
-   * server approval. For approved transfers, this is the actual expiration time of the domain as of
-   * the moment of transfer completion. For rejected or cancelled transfers, this field will be
-   * reset to null.
-   *
-   * <p>Note that even when this field is set, it does not necessarily mean that the post-transfer
-   * domain has a new expiration time. Superuser transfers may not include a bundled 1 year renewal
-   * at all, or even when a renewal is bundled, for a transfer during the autorenew grace period the
-   * bundled renewal simply subsumes the recent autorenewal, resulting in the same expiration time.
-   */
-  // TODO(b/36405140): backfill this field for existing domains to which it should apply.
-  @Column(name = "transfer_registration_expiration_time")
-  DateTime transferredRegistrationExpirationTime;
 
   /**
    * The billing event and poll messages associated with a server-approved transfer.
@@ -115,42 +77,6 @@ public abstract class TransferData<B extends TransferData.Builder<?, ?>> extends
   @Column(name = "transfer_losing_poll_message_id")
   Long losingTransferPollMessageId;
 
-  @Ignore
-  @Column(name = "transfer_billing_cancellation_id")
-  Long billingCancellationId;
-
-  /**
-   * The regular one-time billing event that will be charged for a server-approved transfer.
-   *
-   * <p>This field should be null if there is not currently a pending transfer or if the object
-   * being transferred is not a domain.
-   *
-   * <p>TODO(b/158230654) Remove unused columns for TransferData in Contact table.
-   */
-  @IgnoreSave(IfNull.class)
-  @Column(name = "transfer_billing_event_id")
-  VKey<BillingEvent.OneTime> serverApproveBillingEvent;
-
-  /**
-   * The autorenew billing event that should be associated with this resource after the transfer.
-   *
-   * <p>This field should be null if there is not currently a pending transfer or if the object
-   * being transferred is not a domain.
-   */
-  @IgnoreSave(IfNull.class)
-  @Column(name = "transfer_billing_recurrence_id")
-  VKey<BillingEvent.Recurring> serverApproveAutorenewEvent;
-
-  /**
-   * The autorenew poll message that should be associated with this resource after the transfer.
-   *
-   * <p>This field should be null if there is not currently a pending transfer or if the object
-   * being transferred is not a domain.
-   */
-  @IgnoreSave(IfNull.class)
-  @Column(name = "transfer_autorenew_poll_message_id")
-  VKey<PollMessage.Autorenew> serverApproveAutorenewPollMessage;
-
   public abstract boolean isEmpty();
 
   @Nullable
@@ -158,32 +84,8 @@ public abstract class TransferData<B extends TransferData.Builder<?, ?>> extends
     return transferRequestTrid;
   }
 
-  public Period getTransferPeriod() {
-    return transferPeriod;
-  }
-
-  @Nullable
-  public DateTime getTransferredRegistrationExpirationTime() {
-    return transferredRegistrationExpirationTime;
-  }
-
   public ImmutableSet<VKey<? extends TransferServerApproveEntity>> getServerApproveEntities() {
     return nullToEmptyImmutableCopy(serverApproveEntities);
-  }
-
-  @Nullable
-  public VKey<BillingEvent.OneTime> getServerApproveBillingEvent() {
-    return serverApproveBillingEvent;
-  }
-
-  @Nullable
-  public VKey<BillingEvent.Recurring> getServerApproveAutorenewEvent() {
-    return serverApproveAutorenewEvent;
-  }
-
-  @Nullable
-  public VKey<PollMessage.Autorenew> getServerApproveAutorenewPollMessage() {
-    return serverApproveAutorenewPollMessage;
   }
 
   @Override
@@ -205,9 +107,8 @@ public abstract class TransferData<B extends TransferData.Builder<?, ?>> extends
    */
   public B copyConstantFieldsToBuilder() {
     B newBuilder = new TypeInstantiator<B>(getClass()) {}.instantiate();
-    ;
     newBuilder
-        .setTransferPeriod(this.transferPeriod)
+        // .setTransferPeriod(this.transferPeriod)
         .setTransferRequestTrid(this.transferRequestTrid)
         .setTransferRequestTime(this.transferRequestTime)
         .setGainingClientId(this.gainingClientId)
@@ -232,37 +133,9 @@ public abstract class TransferData<B extends TransferData.Builder<?, ?>> extends
       return thisCastToDerived();
     }
 
-    public B setTransferPeriod(Period transferPeriod) {
-      getInstance().transferPeriod = transferPeriod;
-      return thisCastToDerived();
-    }
-
-    public B setTransferredRegistrationExpirationTime(
-        DateTime transferredRegistrationExpirationTime) {
-      getInstance().transferredRegistrationExpirationTime = transferredRegistrationExpirationTime;
-      return thisCastToDerived();
-    }
-
     public B setServerApproveEntities(
         ImmutableSet<VKey<? extends TransferServerApproveEntity>> serverApproveEntities) {
       getInstance().serverApproveEntities = serverApproveEntities;
-      return thisCastToDerived();
-    }
-
-    public B setServerApproveBillingEvent(VKey<BillingEvent.OneTime> serverApproveBillingEvent) {
-      getInstance().serverApproveBillingEvent = serverApproveBillingEvent;
-      return thisCastToDerived();
-    }
-
-    public B setServerApproveAutorenewEvent(
-        VKey<BillingEvent.Recurring> serverApproveAutorenewEvent) {
-      getInstance().serverApproveAutorenewEvent = serverApproveAutorenewEvent;
-      return thisCastToDerived();
-    }
-
-    public B setServerApproveAutorenewPollMessage(
-        VKey<PollMessage.Autorenew> serverApproveAutorenewPollMessage) {
-      getInstance().serverApproveAutorenewPollMessage = serverApproveAutorenewPollMessage;
       return thisCastToDerived();
     }
