+++ core/src/test/resources/google/registry/model/schema.txt
@@ -716,18 +716,13 @@ class google.registry.model.tmch.TmchCrl {
   org.joda.time.DateTime updated;
 }
 class google.registry.model.transfer.ContactTransferData {
-  google.registry.model.domain.Period transferPeriod;
   google.registry.model.eppcommon.Trid transferRequestTrid;
   google.registry.model.transfer.TransferStatus transferStatus;
-  google.registry.persistence.VKey<google.registry.model.billing.BillingEvent$OneTime> serverApproveBillingEvent;
-  google.registry.persistence.VKey<google.registry.model.billing.BillingEvent$Recurring> serverApproveAutorenewEvent;
-  google.registry.persistence.VKey<google.registry.model.poll.PollMessage$Autorenew> serverApproveAutorenewPollMessage;
   java.lang.String gainingClientId;
   java.lang.String losingClientId;
   java.util.Set<google.registry.persistence.VKey<? extends google.registry.model.transfer.TransferData$TransferServerApproveEntity>> serverApproveEntities;
   org.joda.time.DateTime pendingTransferExpirationTime;
   org.joda.time.DateTime transferRequestTime;
-  org.joda.time.DateTime transferredRegistrationExpirationTime;
 }
 class google.registry.model.transfer.DomainTransferData {
   google.registry.model.domain.Period transferPeriod;
