+++ db/src/main/resources/sql/schema/db-schema.sql.generated
@@ -118,17 +118,10 @@
         addr_local_org text,
         addr_local_type text,
         search_name text,
-        transfer_billing_cancellation_id int8,
         transfer_gaining_poll_message_id int8,
         transfer_losing_poll_message_id int8,
-        transfer_billing_recurrence_id int8,
-        transfer_autorenew_poll_message_id int8,
-        transfer_billing_event_id int8,
-        transfer_renew_period_unit text,
-        transfer_renew_period_value int4,
         transfer_client_txn_id text,
         transfer_server_txn_id text,
-        transfer_registration_expiration_time timestamptz,
         transfer_gaining_registrar_id text,
         transfer_losing_registrar_id text,
         transfer_pending_expiration_time timestamptz,
@@ -182,16 +175,16 @@
         tech_contact text,
         tld text,
         transfer_billing_cancellation_id int8,
-        transfer_gaining_poll_message_id int8,
-        transfer_losing_poll_message_id int8,
         transfer_billing_recurrence_id int8,
         transfer_autorenew_poll_message_id int8,
         transfer_billing_event_id int8,
         transfer_renew_period_unit text,
         transfer_renew_period_value int4,
+        transfer_registration_expiration_time timestamptz,
+        transfer_gaining_poll_message_id int8,
+        transfer_losing_poll_message_id int8,
         transfer_client_txn_id text,
         transfer_server_txn_id text,
-        transfer_registration_expiration_time timestamptz,
         transfer_gaining_registrar_id text,
         transfer_losing_registrar_id text,
         transfer_pending_expiration_time timestamptz,
