+++ db/src/main/resources/sql/schema/nomulus.golden.sql
@@ -244,15 +244,8 @@ CREATE TABLE public."Contact" (
     voice_phone_number text,
     transfer_gaining_poll_message_id bigint,
     transfer_losing_poll_message_id bigint,
-    transfer_billing_cancellation_id bigint,
-    transfer_billing_event_id bigint,
-    transfer_billing_recurrence_id bigint,
-    transfer_autorenew_poll_message_id bigint,
-    transfer_renew_period_unit text,
-    transfer_renew_period_value integer,
     transfer_client_txn_id text,
     transfer_server_txn_id text,
-    transfer_registration_expiration_time timestamp with time zone,
     transfer_gaining_registrar_id text,
     transfer_losing_registrar_id text,
     transfer_pending_expiration_time timestamp with time zone,
